+++ src/freeclimb/Model/AddToConferenceAllOf.cs
@@ -50,7 +50,8 @@ namespace freeclimb.Model
         /// <param name="notificationUrl">When the Participant enters the Conference, this URL will be invoked using an HTTP POST request with the standard request parameters..</param>
         /// <param name="startConfOnEnter">Flag that indicates whether a Conference starts upon entry of this particular Participant. This is usually set to &#x60;true&#x60; for moderators and &#x60;false&#x60; for all other Participants..</param>
         /// <param name="talk">If &#x60;true&#x60;, the Participant joins the Conference with talk privileges. This may be modified later via the REST API or &#x60;SetTalk&#x60; PerCL command. .</param>
-        public AddToConferenceAllOf(bool allowCallControl = default(bool), string callControlSequence = default(string), string callControlUrl = default(string), string conferenceId = default(string), string leaveConferenceUrl = default(string), bool listen = default(bool), string notificationUrl = default(string), bool startConfOnEnter = default(bool), bool talk = default(bool))
+        /// <param name="dtmfPassThrough">If &#x60;true&#x60;, the Participant joins the Conference with dtmfPassThrough privileges. This may be modified later via the REST API or &#x60;SetDTMFPassThrough&#x60; PerCL command. .</param>
+        public AddToConferenceAllOf(bool allowCallControl = default(bool), string callControlSequence = default(string), string callControlUrl = default(string), string conferenceId = default(string), string leaveConferenceUrl = default(string), bool listen = default(bool), string notificationUrl = default(string), bool startConfOnEnter = default(bool), bool talk = default(bool), bool dtmfPassThrough = default(bool))
         {
             // to ensure "conferenceId" is required (not null)
             if (conferenceId == null) {
@@ -65,6 +66,7 @@ namespace freeclimb.Model
             this.NotificationUrl = notificationUrl;
             this.StartConfOnEnter = startConfOnEnter;
             this.Talk = talk;
+            this.DtmfPassThrough = dtmfPassThrough;
         }
 
         /// <summary>
@@ -131,6 +133,13 @@ namespace freeclimb.Model
         public bool Talk { get; set; }
 
         /// <summary>
+        /// If &#x60;true&#x60;, the Participant joins the Conference with dtmfPassThrough privileges. This may be modified later via the REST API or &#x60;SetDTMFPassThrough&#x60; PerCL command. 
+        /// </summary>
+        /// <value>If &#x60;true&#x60;, the Participant joins the Conference with dtmfPassThrough privileges. This may be modified later via the REST API or &#x60;SetDTMFPassThrough&#x60; PerCL command. </value>
+        [DataMember(Name = "dtmfPassThrough", EmitDefaultValue = true)]
+        public bool DtmfPassThrough { get; set; }
+
+        /// <summary>
         /// Returns the string presentation of the object
         /// </summary>
         /// <returns>String presentation of the object</returns>
@@ -147,6 +156,7 @@ namespace freeclimb.Model
             sb.Append("  NotificationUrl: ").Append(NotificationUrl).Append("\n");
             sb.Append("  StartConfOnEnter: ").Append(StartConfOnEnter).Append("\n");
             sb.Append("  Talk: ").Append(Talk).Append("\n");
+            sb.Append("  DtmfPassThrough: ").Append(DtmfPassThrough).Append("\n");
             sb.Append("}\n");
             return sb.ToString();
         }
@@ -176,6 +186,7 @@ namespace freeclimb.Model
             props.Add("notificationUrl", NotificationUrl);          
             props.Add("startConfOnEnter", StartConfOnEnter);          
             props.Add("talk", Talk);          
+            props.Add("dtmfPassThrough", DtmfPassThrough);          
             return props;
         }
         
@@ -241,6 +252,10 @@ namespace freeclimb.Model
                 (
                     this.Talk == input.Talk ||
                     this.Talk.Equals(input.Talk)
+                ) && 
+                (
+                    this.DtmfPassThrough == input.DtmfPassThrough ||
+                    this.DtmfPassThrough.Equals(input.DtmfPassThrough)
                 );
         }
 
@@ -277,6 +292,7 @@ namespace freeclimb.Model
                 }
                 hashCode = (hashCode * 59) + this.StartConfOnEnter.GetHashCode();
                 hashCode = (hashCode * 59) + this.Talk.GetHashCode();
+                hashCode = (hashCode * 59) + this.DtmfPassThrough.GetHashCode();
                 return hashCode;
             }
         }